fix warnings
git-svn-id: https://advants.svn.sourceforge.net/svnroot/advants@196 fd2da0a4-d968-48df-9cb5-d8c496287cbf

+++ ImageRegistration/ANTS_affine_registration2.h
@@ -500,7 +500,7 @@ void InjectInitialPara(PointType &center, VectorType &translation_vec, Transform
     ParaType para0(TransformPointerType::ObjectType::ParametersDimension);
 
 
-    switch(PointType::PointDimension){
+    switch((unsigned int) PointType::PointDimension){
     case 2:
         para0[0] = 0; // para1[0]; // theta
         para0[1] = 1.0; //s1
